Updated mediawiki/extensions/Wikidata Project: mediawiki/extensions/Wikibase  7d1437ae45d56f021ee6e7f12caef416388a4d02
(bug 58344) use getUserLang for EntityView

When generating page output, we should consistently use the
language provided by ParserOptions::getUserLang, and not
rely on the language provided by the context language.

This change also kills some dead wood in related code paths,
and fixes improves tests.

DEPLOYMENT: purge the parser cache by bumping $wgCacheEpoch.
That was really already required by the change that enabled
the parser cache for entity pages, but is mentioned again here,
so we don't forget to do it.

Change-Id: I77cfdb20eccbf4c15189b925fc4190e2806c6b2c