Updated mediawiki/extensions/Wikidata Project: mediawiki/extensions/Wikibase  80edaca54dab9700307f21cd8ccc4e5fd4333374
Use correct EntityIdParser interface and implementations.

There was a lot of confusion in the code base about Wikibase\Lib\EntityIdParser
versus Wikibase\DataModel\Entity\EntityIdParser. This change makes sure that
the two have different names, and the correct one is used throughout the code.

There are still a lot of places where proper injection of the ID parser is missing.

Change-Id: I276537c592cf229ac0877e67151c0b4a01421994