Use default wiki id in example settings.
Until now, we force the local wiki ID to "mywiki" in the exampel settings.
This triggers a warning if "mywiki" is not known to the SiteLookup.

Instead of forcing a fixed value, we can just rely on the default for the
local site ID, which is the wiki's database name. That is consistent with
the conventions used on the WMF cluster.

Cherry picked from https://gerrit.wikimedia.org/r/#/c/329765/1

Bug: T153729
Change-Id: I01966cfb24a0700541a1d93c6f1ce5fd639b8c10